cant loose bosses now

+++ interactBDD.py
@@ -611,7 +611,7 @@ class InteractBDD(Static):
 
 			
 			[conn2, cur2]=InteractBDD.beginQuery()
-			request = "SELECT fruit FROM pirate WHERE username='"+username+"';"
+			request = "SELECT fruit FROM pirate WHERE username='"+username+"' AND name NOT IN (SELECT nom FROM pnj);"
 			description = InteractBDD.connectAndExecuteRequest(request, False, conn, cur)
 			for elem in description:
 				fruitsName=str(elem[0]) # we got the name of the current archipel
@@ -619,7 +619,10 @@ class InteractBDD(Static):
 				InteractBDD.connectAndExecuteRequest(request, True, conn2, cur2)
 			InteractBDD.endQuery(conn2, cur2)
 
-			request = "DELETE FROM pirate WHERE username='"+username+"';"
+			request = "DELETE FROM pirate WHERE username='"+username+"' AND name NOT IN (SELECT nom FROM pnj);"
+			InteractBDD.connectAndExecuteRequest(request, True, conn, cur) 
+
+			request = "UPDATE pirate SET level=1 WHERE username='"+username+"';"
 			InteractBDD.connectAndExecuteRequest(request, True, conn, cur) 
 
 			InteractBDD.endQuery(conn, cur)
@@ -631,7 +634,7 @@ class InteractBDD(Static):
 			[conn, cur]=InteractBDD.beginQuery()
 			
 			[conn2, cur2]=InteractBDD.beginQuery()
-			request = "SELECT fruit FROM pirate WHERE username='"+username+"';"
+			request = "SELECT fruit FROM pirate WHERE username='"+username+"' AND name NOT IN (SELECT nom FROM pnj);"
 			description = InteractBDD.connectAndExecuteRequest(request, False, conn, cur)
 			for elem in description:
 				fruitsName=str(elem[0]) # we got the name of the current archipel
@@ -639,8 +642,11 @@ class InteractBDD(Static):
 				InteractBDD.connectAndExecuteRequest(request, True, conn2, cur2)
 			InteractBDD.endQuery(conn2, cur2)
 
-			request = "DELETE FROM pirate WHERE username='"+username+"';"
+			request = "DELETE FROM pirate WHERE username='"+username+"' AND name NOT IN (SELECT nom FROM pnj);"
 			InteractBDD.connectAndExecuteRequest(request, True, conn, cur)
+
+			request = "UPDATE pirate SET level=1 WHERE username='"+username+"';"
+			InteractBDD.connectAndExecuteRequest(request, True, conn, cur) 
 
 			InteractBDD.endQuery(conn, cur)
 			return None
@@ -651,7 +657,6 @@ class InteractBDD(Static):
 				crew=obj.equipage
 			else: #equipage
 				crew=obj
-
 
 			[conn2, cur2]=InteractBDD.beginQuery()
 			for pirate in crew.team:
@@ -681,17 +686,33 @@ class InteractBDD(Static):
 		def removeFighter(username, pirate):
 			[conn, cur]=InteractBDD.beginQuery()
 
-			fruitsName=pirate.fruit.name
-			if fruitsName!="None":
-				request = "UPDATE fruit SET allocated=0 WHERE name='"+fruitsName+"';"
+			boss=False
+			request = "SELECT name FROM pnj WHERE name='"+pirate.name+"';"
+			description = InteractBDD.connectAndExecuteRequest(request, False, conn, cur)
+			for elem in description:
+				boss=True
+
+			if boss:
+				
+				request = "UPDATE pirate SET level=level-10 WHERE username='"+username+"' AND name='"+pirate.name+"';"
+				InteractBDD.connectAndExecuteRequest(request, True, conn, cur) 
+
+				InteractBDD.endQuery(conn, cur)
+				return None
+
+			else:
+
+				fruitsName=pirate.fruit.name
+				if fruitsName!="None":
+					request = "UPDATE fruit SET allocated=0 WHERE name='"+fruitsName+"';"
+					InteractBDD.connectAndExecuteRequest(request, True, conn, cur)
+
+
+				request = "DELETE FROM pirate WHERE username='"+username+"' and name='"+pirate.name+"' and fruit='"+pirate.fruit.name+"' and qualite='"+str(pirate.qualite)+"';"
 				InteractBDD.connectAndExecuteRequest(request, True, conn, cur)
 
-
-			request = "DELETE FROM pirate WHERE username='"+username+"' and name='"+pirate.name+"' and fruit='"+pirate.fruit.name+"' and qualite='"+str(pirate.qualite)+"';"
-			InteractBDD.connectAndExecuteRequest(request, True, conn, cur)
-
-			InteractBDD.endQuery(conn, cur)
-			return None
+				InteractBDD.endQuery(conn, cur)
+				return None
 
 
 		#____________________________________________________________
